Sets chain to bitcoin::network::constants::Network
Currently, when reading a chain field from an rpc response from bitcoind the field
is parsed as String. If this is happen to be compared with a `bitcoin::network::constants::Network`
a custom parser would need to be created, given bitcoind uses `main` and `test` for mainnet and testnet
while `rust-bitcoin` uses `bitcoin` and `testnet`.

Creates a custom deserializer based on Network::from_core_arg so both can be compared seamlessly