Add verifymessage feature
In #326 we changed a function to use
`bitcoin::sign_message::MessageSignature` but doing so requires the
"base64" feature to be on for `bitcoin`. This did not get caught by CI
but improvements to CI in #338 will now catch this.

Add a feature to `json` and `client` that allows enabling "base64" if
the `verifymessage` RPC call is required.