START FORK OF CRATE
I am forking this crate because of differing views with the original
author. In particular Steven wants to support async.

I will be providing this fork explicitly for integration testing of
crates in the `rust-bitcoin` eccosystem.

Please be aware that I will be:

- Merging code with no acks.
- Moving fast and breaking things.
- Not guaranteeing anything because basically I have zero interest in
  JSON-RPC I just want to be able to test our code against Bitcoin Core.

PRs welcome, use at your own risk.