CI: Use script from rust-bitcoin-maintainer-tools
We have a CI script in the `rust-bitcoin-maintainer-tools` repository,
lets use it.

Requires adding `Cargo-minimal.lock` and `Cargo-recent.lock`. Both these
lock files are tested with all three toolchains (msrv, stable, nightly).

Note, I could not find a more recent set of dependencies that works so
the two lock files are the same.