Add initial RPC calls using jsonrcp-client
We would like to test the `bitcoincore-rpc-json` crate without
depending on `bitcoincore-rpc` - the point being that users of the
crate should be able to use what ever RCP client they want.

Add infrastructure and a few initial RPC calls to see what it looks
like to test `json` using the `jsonrpc-client` crate.

The `run.sh` script is just a quick copy and hack of the other one here.

The aim is to eventually test the `json` RPC types against specific
versions of Core to make sure we have them correct as Core changes.